added LineSubtractionModule
